requirements.txt
flask
numpy
scipy
pillow
python-dotenv
